Removed setBpm() retriggers looper playback
Although this allows loops to go out of sync, I think its more
"musical" than the re-trigger noises of scrubbing the BPM dial...

In short - in a live performance I do not expect musicians to
change tempo in a song - between songs it possible, and then
triggering a scene will bring everything back into sync.

This needs a better solution, but right now I'm not sure how
to best do that. The other changes in the patches that changed
this behaviour are valuable, so commiting to get closer to a
solution - but I'm not comfortable with the retrigger noises.