midi: allow clearing clip from GenericMIDI binding
This commit allows the user to clear clips from the Luppp grid
with a MIDI message. The binding method is very similar to the
existing "grid:event" style, with added to "action" representing
1 for press, and 0 for release, -1 now represents a "clear" event.

As such, the following binding would clear clip track 1, scene 1,
when midi note number 60 had a note on event. Note that the value
for "send" is also set to -1, indicating it is not a send binding.

Active set to -1 indicates that the clip should be cleared:

{
  "action": "grid:event",
  "status": 144,
  "data":    60,
  "track":    1,
  "scene":    1,
  "send":    -1,
  "active":  -1
}

See Github issue #134