ctlra: v1 implementation imported and cleaned up
This is an import and rework of previous POC of the Ctlra
library and Luppp. The initial code had many rebase conflicts
with Luppp (it was done during Nov 2016, and as a POC for the
TCC scripting components, hence was not good quality and each
commit wasn't well split - hence this single bulk commit).

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>