mappa: initial implementation with track vol/send
This commit is the first commit to support Mappa,
the "mapping" solution from the Ctlra project.

The code here implements two targets as callback
functions, making use of the Mappa "token" when registering
a target. The track-id is stored in the token, allowing easy
identification of the software parameter to tweak.

The main UI thread polls the mappa library, which internally
handles all Ctlra APIs. The application sees its callbacks
being called when a control surface is moved. The callbacks
are invoked in the same thread as that which calls mappa_iter().

The mapping from control-surface to software-target is done using
a simple .ini file, which the user supplies. The format is still
in flux, but the basic structure is not expected to change.

The user provides multiple "layers" of mappings, and Mappa provides
a mechanism to switch between these layers on the fly. This enables
complex multi-layered controller setups, without any application
logic complexity.

Note that the latest commit from the Mapping PR of Ctlra is required
in order to get access to mappa.h, see details of PR here:
https://github.com/openAVproductions/openAV-Ctlra/pull/87

This resolves the following Ctlra issue to do real testing with Luppp:
https://github.com/openAVproductions/openAV-Ctlra/issues/90

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>