controller: add GenericMIDI transport start/stop
This commit adds support for starting to stopping the
transport using the GenericMIDI system. A test .ctlr
file has been added to the resources dir, and the code
to implement loading of start/stop is updated.

Note that now the MidiIO input is read from Jack::process()
even with transport in "stopped" state, while before all
input was ignored unless we were in transport rolling.

This commit relates to #261.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>