Unify of single clips and the hole session
This introduces a new Event, called EventSaveClip which starts saving of a single clip. On the other side, the confusing and not anymore needed no_dealloc option in the
EventStateSaveBuffer was removed. I think this is now easier to understand.

Possible problem: The differentiation of the saving location is done by checking if gui->saveBufferPath is set, feedback here is welcome.